src/services/user/requirements.txt
grpcio
grpcio-tools
asyncpg
pydantic
passlib
bcrypt